commit-ish doesn't include tags today
gh release --target only accepts a branch ref or a FULL sha hash, it doesn't accept tag names, which surprised us

we tried writing a version where we just supply a sha hash instead of creating a tag and then passing the tag name to the workflow, but workflows can't be run on a sha hash, they can only be run on tags and branches.

so we reverted to the original strategy of creating a tag locally and pushing it, and then starting the workflow with that tag, but then not referencing the tag for gh release, hoping that `${{github.sha}}` will work instead.