Change `docs` command to expect `Name`
Previously, when given a numbered arg, from some commands (e.g., `find`), it
would fail to find the docs because the hash associated with the definition was
applied to the `doc`, which then would be incorrect.

This now discards hashes up-front, so it can add the `doc` suffix to the name.

Fixes #3939.