Ensure `Name`s are absolute after prefixing
In a separate part of the change this PR was extracted from,
`Path.prefixName` ensures the result is absolute, but that is more
involved than other changes here, so this uses the previous technique of
calling `Name.makeAbsolute` afterward.

I don’t think this changes anything important, but the `Name` _is_
always absolute after `Path.prefixName`, so this just ensures that it’s
tracked.