Improve CLI shared error message
This rephrases part of it and includes instructions to run the `help`
command (rather than including the `help` output directly). In future,
we may directly include shorter `help` outputs or a separate set of
examples. But for now the `help` output may overshadow the error message
we intend to show the user.