Improve some lexer error messages
The message mentioned in #5060 was incorrect. This also uses the passed-
in `id` for that message and related messages (removing the reliance on
ANSI colors for conveying where the error is). And it adds a suggestion
on how to avoid the error.

It also does some minor adjustment of highlighting – styling individual
identifiers rather that a list of them.

Fixes #5060