Rework define-unison to be based on static 'curry' functions
Directly building closure structures seems to have foiled racket's
optimizer, but it's also desirable to avoid expanding code for all
definitions. So, this pre-generates currying functions and makes
the main definition of each unison function expand into just the
currying of the implementation body.

Right now a fixed number of procedures is produced. A fail-over to
a general procedure in case we have more arguments than that is TODO.