Rewrite runtime code loading to fix some problems
The existing approach had a faulty check. It was supposed
to avoid loading code that was already in the runtime system.
However, it instead avoided loading any code whose dependencies
were entirely satisfied by what the runtime system already has.

The structure of the old loader was a bit weird for doing the
correct thing (due to relative inexperience with scheme at the
time), so I rewrote the overall function to be structured a
bit better.

After fixing that, it became clear that newly loaded code
depending on already loaded code was just unimplemented, so I
implemented it.