Improve transcript runner output
Previously, it would output messages like

    ⚙️   Processing stanza Just 6 of 7.

and

    ⚙️   Processing stanza Nothing of 7.

which was especially confusing when there was text or some other
non-Unison block at the end of the transcript.

Now the messages look like

    ⏩   Skipping non-executable Markdown block.
    ⚙️   Processing stanza 6 of 7.
    ✔️   Completed transcript.

The one shortcoming is that I don’t know how to clear the line after the
carriage return, so I added whitespace padding to make sure the previous
messages get overwritten.