Be more careful about having code vs. loading code
In a compiled binary, we'll _know_ the code for all our
definitions, but it won't necessarily be loaded into the
runtime namespace used for dynamic code loading (due to
racket peculiarities). So, we need to make sure we don't
assume that knowing the code for a definition means that we
have evaluated a module with it defined.