Rework code loading to fix compiled scenarios
Although compiled code is emitted so that the code objects are known,
they are not loaded as accessible procedures immediately. This means
that if you just expect to have certain code (because it's involved in
the program you're actually running), in a compiled executable you may
have been wrong.

This means that the loading builtins need an extra step to detect
dependencies for which code is known, but hasn't been loaded yet. Then
the items to be loaded can be augmented with the known code.

To make this happen, I've refactored the code loading into multiple
procedures. One or two check for actual missing dependencies and signal
appropriate results. One crawls the known-but-unloaded code for
additional items to load. Finally, one actually loads a module that is
complete with respect to a set of code to load. These can be assembled
in various ways to implement Code and Value loading.

Some of the structures involved have been reworked a bit as well, to be
more convenient. Termlinks have been used more systematically when
possible, and racket hash maps are used to store code associations rather
than lists that we `remove-duplicates` on.