Make annotations on Doc more consistent
This does all `Doc` annotations with `Cofree`. As a consequence, some
types have been broken apart and constraints that were mentioned in
comments before are now encoded in the types. The only remaining direct
recursion is now between `Column` and `List`.

This is a precursor for unit testing Doc parsing, as this allows us to
drop all annotations from the Doc structure, making it much less
fragile.