Tweak MCode Let representation
This introduces a minor change to the structure of MCode Let
bindings. Before various refactoring, a Let would use a CombIx to
indicate where execution should resume once the binding is
evaluated. Now the code is resolved ahead of time and referred to
directly. However, logically this code is just a further Section;
the rest of the actual function that contained the Let.

The purpose of using a Comb was to store stack protection
information, so that the stack check can be performed upon reentry.
This is necessary if the continuation has been captured and we are
resuming in a context that was not the original entry point of the
resumption.

Now, though, a top level 'comb' can be either code or a memoized
value, but a resumption will never be the latter. So Let has been
changed to store the relevant information, instead of delegating to
Comb.