Make transcript output valid Markdown
All serialization is now down via CMark (previously, all `api` and `ucm` blocks in the output were still being
serialized directly, because of the interleaving of the input text and command/request output).

This causes a few changes to the existing transcript outputs:
- there is now always a blank line between Markdown block elements,
- leading blank lines in UCM blocks are gone,
- blank lines at the end of transcripts are gone (they still end with a final POSIX newline), and
- some `ucm` blocks now have 4-backtick fences (because they contain triple-backticks).

Transcript failures are now also handled as Markdown, rather than just being text appended to the document. This mostly
doesn’t change the serialization, except that the failure message is now fenced, since they often contain newlines and
indentation that is mangled otherwise.