Add an `:incorrect` transcript tag
This is basically the opposite of `:failure` – it indicates that a
successful result is incorrect. The correct result may either be an
error (which will be caught by the transcript runner) or a different
result then what is currently expected (which won’t be caught by the
runner, but the appearance of `:incorrect` conveys to the programmer
that the output.md diff may be an improvement from the previous state).