Change the value serialization format to be an extension of v4
Instead of serializing code as:

  CodeTag|SuperGroup|Cacheability

use one of the two following:

  CodeTag|SuperGroup        <-- uncacheable
  CacheCodeTag|SuperGroup   <-- cacheable

Since this is a strict addition to the format, it's not necessary to
bump the version to properly parse newly serialized values. Old
installations will fail to recognize certain byte representations,
but in a way that they are able to recognize and error on, rather than
silently parsing things incorrectly.