Update to new unison internal libraries, with support
I'd already been working on caching values in the jit in the
internal libraries, so it was convenient to include those changes in
the update that also parses the extended v4 format. This requires a
bit of additional racket code, but the consequences are disabled
until some supporting code is available in the base library.

Some definitions will be tagged as `value` indicating that they can
be expanded into racket definitions that will only be evaluated
once. The code for expanding things that way is included, too. But
the feature is turned off, so everything will expand in the old way.