Re-enable automatic runs of Nix cache workflow
Instead of doing it for every PR push, only run it for trunk & releases.
The former keeps an up-to-date dev environment in the cache (and lets us
know if something needs to be fixed before the next release) while the
latter gives us something to pin for downstream consumers (and quick
access to releases when bisecting for bugs, etc.).