Fix the intentionally mis-formatted code manually
The Ormolu job apparently doesn’t work on forks or protected
branches[^1], so have it do `check` instead of `inplace` in those cases.

[^1]: I’m a bit confused by this, because when I put up an earlier
version of the PR, it _did_ create a commit (which only formatted one of
the files), but then I couldn’t get it to do the same on subsequent
pushes.