Roll back immediate affine handler compilation
This keeps the affine handler spot in the intermediate code
representation, but it is not generated during compilation, and is
omitted from the interchange format. The idea is to generate the
affine handlers as a post-process around the time inlining happens.
This should make it easier to recognize affine handlers, and also
means that code interchange, hashes, etc. are unaffected by this
optimization.