Add ANF support for affine handlers
These are not generated yet, but they are the machinery necessary for
the optimized handlers. As a first step, the top level exception
handler has been changed to an affine handler.

These constructors are also not part of the serialization format. They
will be generated when the code is loaded so that hashes etc. do not
need to change to support affine handlers.