Avoid unnecessary evaluation of handler environments
The handler state is not used everywhere, so it's not really
advantageous to keep it evaluated so that the `HEnv` value is unpacked
as arguments to `eval`, `exec`, etc. Instead, the values are forced
when built (to avoid excessive thunking), and inspections of them are
performed only inside cases where they're necessary, to avoid needless
forcing operations.