Run Ormolu over the entire repo
This shows the shortcomings of the previous push-based workflow. In
bee1f7a, the push-based run was triggered, but it only ran on the files
that were updated by the single merge commit prior to the automatic run.
It didn’t affect files that were changed by other commits in the PR
(like Stack.hs, which is updated here and caused the pull-request based
check to fail).