Fix a problem with floating w/r/t some enclosed terms
An enclosed lambda term with an annotation ends up looking like:

  w x -> (y z -> ...) : ...

This is actually a single lambda term; the inner, annotated term
is not necessarily closed. However, floating was handling the
lambda terms separately, floating the inner, open term as well as
the outer term. The exception is if the overall term is let-bound,
which was handled as a single term properly.

This latter case is the more common occurrence, because it is the
result of annotated bindings. But if you manually annotate a
lambda term, you could trigger the other case, which resulted in
exceptions during rehashing.