Fix some issues causing cli integration test failures
Optimization info for builtin stuff wasn't getting stripped out. This
was causing the top level exception handler to be written out, and it
contains a `Discard` which we refuse to serialize.

---

Serializing optimization info involves inlined foreign functions,
which we previously refused to read back in. I've changed that, and
the scheme for serializing them.

They were never part of the public interchange format. The only reason
for supporting their serialization was for the scheme compiler. But
now it directly implements builtins, so it shouldn't need to know
about the wrappers, which is the only place they would occur.

I've switched to using numbers for foreign functions. It uses the
derived enum numbering, because this is just for standalone programs,
which don't have stable versioning (each new ucm release is a separate
compiled version). There is a flag passed to the serialization which
controls whether a foreign op occurrence causes an error. The error
should occur on all interchange level stuff, and only not occur for
the standalone serialization.