Fix a bug in one of the peephole optimizations
The HandlerResume case was not checking that the lazy binding it was
eliminating was actually unused except for being forced in a handle
block. _Now_ I think it is basically impossible to write such code,
because the erroneous case was caused by failing to cover all the
cases in an ability matching block. However, there was an example in
an older version of `codec` that was written before exhaustiveness
checking was implemented.

I've fixed the optimization to check the condition even though it
shouldn't happen on any newly written code.