Relax recognition of handler entries for affine optimization
Previously, the optimization was assuming that a handler always looks
like:

    let rec
      matcher <vs> rq = ...
    in
      <vs> rq ->
        h = matcher <vs>
        handle !rq
        with h

where `<vs>` is a list of variables. Notably, the variables were
expected to be in exactly the same order.

However, it turns out that the variables can be permuted, because the
way floating of each works is slightly different. Essentially, the
body (the top level entry point) comes with an argument list, and
additional free variables are prepended, but `matcher` just gets its
variable list generated wholly from the set of free variables.

I don't think it makes any difference if the variable lists are
exactly the same, or even if the matcher uses all the variables of the
entry point. So I have removed that condition.

This allows a handler in @unison/codec to be recognized correctly. The
situation arises when a locally defined handler references free
variables.