Improve optimizer to recognize more affine handlers
This refactors things a bit to be able to apply extra optimizations
before trying to recognize affine handlers. The extra optimizations
are only safe in an affine context, because they may rearrange values
on the stack, or inline expressions in ways that won't be properly
reconstructible on serialization. However, these differences don't
matter, because stacks are not captured in affine contexts.

The improvements allow recognizing affine handlers that are
significantly more complicated than the most obvious examples. For
instance, previously, if a 'nice' affine handler were written such
that the outermost definition did anything besides immediately
handling the forced thunk, it basically wouldn't work. That is no
longer a restriction.