Improve caught runtime panics in tryEval
Previously it was falling through to the generic case and being run
through `show`. However, we can extract a message and a value and
build a better unison exception with those.

This means we can get a 'pretty printed' version of the value in
situations where the panic reports one, rather than showing the raw
structure, which can be a lot worse.