Use lazy bytestring for Value deserialization
The old version was using `toArray` to turn the unison bytes into a
form that can be consumed by the serializer. This seems to perform
quite badly. I think any concrete choice would be better, but I chose
lazy bytestring because it seems slightly better to produce a chunked
input from the rope that is already chunked.