Don’t sandbox `unison-cli:test:cli-tests` in Nix
On macOS, it requires access to `security`, which is outside the
sandbox, so don’t sandbox it when running on Darwin.

This is made possible by setting `sandbox` to `relaxed` in the flake,
which allows using `__noChroot` to mark derivations as unsandboxed. This
makes things less strict on Linux (where the sandbox is on by default),
but more strict on Darwin (where the sandbox is off by default).