Print `Doc` terms as `Doc2`
This helps with migration, as any existing `Doc` values will be inserted
into scratch files as `Doc2` values, which can then be parsed as such.

This doesn’t do a perfect translation. E.g., `Doc` and `Doc2` have
different notions of “paragraph”, and so the `Doc2`-formatted output has
the same newlines as the original `Doc`, which may not be sufficient for
maintaining distinct paragraphs when parsed as `Doc2`.