Add a special case for maps in the new serialization format
Intermediate values now just store the `toList` of a map. The old
serializer handles this by expanding in-situ. The new one just
serializes the list of pairs. This makes sending maps in the new
format considerably faster, while the overall costs are comparable for
the old format, just moved around.