Tweak Cont representation and add Reference traversal functions
Cont was storing a Map for some reason, but there isn't much reason
for it. Switched to a list of pairs, with corresponding serializer
changes. The format is not any different, since maps were represented
as lists of pairs there.

Reference traversal functions traverse all the References in a Value,
Cont, etc. These are distinct from some of the traversals/folds that
were there before, because those traverse references needed for
loading.