Refactor some intermediate and serialization code
This contains a few related changes.

The code for canonicalizing the references in Code values was
basically a duplicate of the one from the Referenced module. It has
been replaced with that version.

The ANF module was re-exporting Referenced stuff. That has been
removed, and the downstream modules now import Referenced.

The intermediate representation of request matching has been switched
to a list of pairs. The reason for this was that `Map.fromList` gets
optimized in a way that rebuilds some `Reference` keys. This foils the
in-memory canonicity of the references, and causes crashes during
serialization. The association list doesn't have this problem.

This is more of a stop-gap fix for the problem while a more robust
solution not relying on `StableName` can be written.

  #	modified:   unison-runtime/src/Unison/Runtime/Machine/Types.hs
  #	modified:   unison-runtime/src/Unison/Runtime/Referenced.hs