Implement customizable transcript behavior
This adds a set of behaviors (currently `autoupdate` and `getHidden`) as
well as a transcript `type` that allows setting multiple behaviors at
once (which can then be overridden).

This also adds a `:show`  info tag, which can be used to override the
default `:hide` that’s enabled by the tutorial type.