Improve an error message when an ability variable can't be solved.
In definitions like:

  foo : '{g} a -> ()
  foo th =
    !th
    foo th

we end up in a situation where we can't actually solve to the type
`'{g} a ->{g} ()`. The reason is that we make up the existential
variable immediately so that we can type check the recursive function
with a specified type of `'{g} a ->{e} ()`. The `g` that `e` needs to
be solved to ends up later in the context, so the solution is not
allowed.

This change introduces a new error message that detects this situation
and suggests an ability set for the signature. There was a previous
'pretty' error message that had been disabled for this situation, but
it was fairly uninformative, and might have been just as useless as
the ugly error message we were currently displaying.