Remove 1 of 2 remaining `ORMOLU_DISABLE` pragmas
The other one is more entangled, but this one just required breaking out
some derived instances.

Not sure if this is a win, but I had the sitting in a stash, so figured
I’d let someone else decide.

It also turns the comments on the affected `data` decl into Haddock.