Handle `Arr` and `Map` cases in `blitLinks`
This is the function that folds over links necessary to load a value. It
wasn't handling the case for arrays or the new case for maps, so it was
possible to miss dependencies for those sorts of values. Then trying to load
the value would fail if the runtime cache wasn't already primed with the
missing dependencies.