Add `check-weeds` script
This script also gets run as part of `scripts/check.sh`.

Since `check-weeds` has to do much of the same work as
`check-formatting`, I extracted the common code to a separate Bash file.

Also made it so the scripts don’t have to be run from the repo root
directory.