Fix `install-hooks.bash` script
I hadn’t realized there were Git hooks, but noticed them when I was
adding the Weeder script.

So, I documented that they exist, and then fixed the script since it
wasn’t working for me.

It now works
- when run it from anywhere in the repo (not just from two directories
  down somewhere),
- when `core.hooksPath` is set (i.e., `$GIT_DIR/hooks` isn’t the right
  place),
- when run from a worktree (which doesn’t necessarily have its own hooks
  dir),
- when the hooks directory doesn’t exist, and
- when the hooks already exist (if you pass `-f`).

This also replaces `ln -s` with `cp` so that in case the particular
worktree it was run from is renamed or deleted, it doesn’t break the
hooks for the repository.