Correct `Semi`s when reordering stanzas
We pull types, ablities, and `use` to the beginning of files and
definitions. But if they occur at the end of a block, they don’t have a
trailing `Semi`. Reordering in that case would create an invalid stanza.

This ensures we add trailing `Semi`s to any stanza not at the end of a
block.

Fixes #686.