Fix `workflow_dispatch` trigger for Cabal workflow
The changes to support commenting on a PR broke the ability to manually
run the workflow. This condionalizes the parts required for
`pull_request_target` so that running on a branch explicitly will work.

It also conditionalizes the `continue-on-error` value so that runs
triggered by `workflow_dispatch` will fail (if something goes wrong)
rather than hiding the failure as we do when we add a comment to the PR.