Make the Cabal workflow cache more precise
It was previously only based on the cabal.project.freeze file, which
changes very rarely. However, a change to build flags (as happened
in #5856) can cause dependencies to need to be rebuilt. But if the cache
key was a perfect match, we wouldn’t cache the updated dependency
builds.

This adds a hash for all Cabal files to the key, so any change that
would affect dependencies causes a cache miss. It also adds the old key
(which is a prefix of the new one) as a restore key, so even when cabal
files change, we load a cache for the correct LTS (GHC version, etc.)
which can prevent cache churn in the period after LTS updates.