Implement a custom binary getter
Uses a fixed strict ByteString for the input and a PrimVar for
the current index, and primitive monads (ST or IO) to do
operations and yield results. This allows the parser to just
return the answers, which seems to allow for better overall
optimization than plumbing the parse state through in a
State-like manner. In the latter, it is very sensitive to get the
compiler to actually unbox all the results/state returned
reliably.

The downside is that the entire byte buffer must be in memory
(since it's a strict bytestring). But that probably isn't a big
deal for serialized unison values.