Fix a bug with continuation calls
When continuations with captured pending arguments were called, it
was possible to have situations that discarded those arguments,
similar to an erroneous `yield`. There is explicit logic in the latter
to apply the yielded result to the pending arguments that was absent
in the continuation case. This adds comparable logic to the
continuation call.