Replace type checking Context with a FingerTree based structure
The new representation allows for splitting and rejoining contexts more
efficiently. The `Info` type that used to be built up linearly,
left-to-right, is now aggregated monoidally (and implicitly) from the
elements stored in the tree. The info allows for efficiently splitting
contexts at a specific variable (necessary for type checking), and
preserves aggregated info in segments. By not discarding this aggregate
info, it can just be merged back together when the overall context is
rebuilt, which saves linear amounts of work when e.g. solving variables.