Lift some file creation toward `main`
Deep file creation was causing tests to modify files in `$HOME`, which was exposed by Nix sandboxing. This moves the
file creation out a bit so that test code can specify an alternative location (alongside the test codebase).

I’m not sure this is the right level – should we just not manage credentials during a test? Have a null manager? And
this really should be moved up even further.