block for lock on second codebase open
the transcript.inplace runner seems to open the codebase for migrations,
then opens it a second time to tun the actual transcript. there seems to
be a race condition in which releasing the lock the first time doesn't
actually block for the lock to be released; and in fact it doesn't seem
to be released in time to acquire the lock the second time, and we were
seeing crashes in CI as a result.

there were a few other places where it looked like it might be fine and
good (though not necessary) to also block for the lock, but i left them
out in order to have a more surgical PR.